Makes possible to set where the horizontal or/and vertical labels should be placed
This is the second step of changing the GraphView into a RealtiveLayout
instead of a LinearLayout.
The horizontal labels are not longer drawn as part of the graphContent
but externalized on their own view. Same goes for the title so we can
define a custom view instead of just text.
To change horizontal and/or vertical labels position set
GraphViewConfig.HORIZONTAL_LABEL_RELATIVE_POSITION and/or
GraphViewConfig.VERTICAL_LABEL_RELATIVE_POSITION correspondingly.