Added ability force bar graph width
Our use case is a 2 hour viewport of data. We know each datapoint is a
minute in size. Therefore we can force the bar width to be 60,000ms.
This way when the first few measurements are received, we dont show the
first point as 1/2 the graph, the next as 1/4 graph width etc.